app/main/services/hope_net/requirements.txt
Flask==1.1.4
flask-restx==0.4.0
numpy>=1.18.1
opencv-python
Pillow==7.0.0
torch>=1.4.0
torchvision>=0.5.0